+++ Yamama/Repository/IProduct.cs
@@ -9,7 +9,7 @@ namespace Yamama.Repository
    public  interface IProduct
     {
         //to get the price for specific product(this function will be called in the balance service)
-        Task<decimal> GetProductPrice(string name);
+        Task<Double> GetProductPrice(string name);
 
         //to get specific product details(this function will be called in the balance service)
         Task<Product> GetProduct(string name);
