+++ Yamama/ViewModels/TransporterReports.cs
@@ -9,7 +9,7 @@ namespace Yamama.ViewModels
     {
         public string product { get; set; }
         public string client { get; set; }
-        public int Qty { get; set; }
+        public Double Qty { get; set; }
         public DateTime date { get; set; }
         public string TransporterClient { get; set; }
         public int ID_invoice { get; set; }
